C3: Pass C3Context directly to generator functions
