+++ packages/create-cloudflare/src/workers.ts
@@ -1,6 +1,6 @@
 import { cp, mkdtemp, readdir, rename, rm } from "fs/promises";
 import { tmpdir } from "os";
-import { dirname, join, resolve } from "path";
+import { join, resolve } from "path";
 import { chdir } from "process";
 import { endSection, startSection, updateStatus } from "@cloudflare/cli";
 import { processArgument } from "@cloudflare/cli/args";
@@ -17,32 +17,17 @@ import { detectPackageManager } from "helpers/packages";
 import {
 	chooseAccount,
 	gitCommit,
-	isInsideGitRepo,
 	offerGit,
 	offerToDeploy,
 	printSummary,
 	runDeploy,
-	setupProjectDirectory,
 } from "./common";
-import type { C3Args, C3Context as Context } from "types";
+import type { C3Context } from "types";
 
 const { dlx } = detectPackageManager();
 
-export const runWorkersGenerator = async (args: C3Args) => {
-	const originalCWD = process.cwd();
-	const { name, path } = setupProjectDirectory(args);
-
-	const ctx: Context = {
-		project: { name, path },
-		args,
-		originalCWD,
-		gitRepoAlreadyExisted: await isInsideGitRepo(dirname(path)),
-	};
-	const { preexisting } = await getTemplate(ctx);
-
-	if (preexisting) {
-		ctx.args.ts = false;
-	}
+export const runWorkersGenerator = async (ctx: C3Context) => {
+	ctx.type = "workers";
 
 	ctx.args.ts = await processArgument<boolean>(ctx.args, "ts", {
 		type: "confirm",
@@ -73,7 +58,7 @@ export const runWorkersGenerator = async (args: C3Args) => {
 	await printSummary(ctx);
 };
 
-async function getTemplate(ctx: Context) {
+async function getTemplate(ctx: C3Context) {
 	const preexisting = ctx.args.type === "pre-existing";
 	const template = preexisting ? "hello-world" : ctx.args.type;
 	const path = resolve(
@@ -88,7 +73,7 @@ async function getTemplate(ctx: Context) {
 	return { preexisting, template, path };
 }
 
-async function copyFiles(ctx: Context) {
+async function copyFiles(ctx: C3Context) {
 	const { template, path: srcdir } = await getTemplate(ctx);
 	const destdir = ctx.project.path;
 
@@ -100,7 +85,7 @@ async function copyFiles(ctx: Context) {
 	await rename(join(destdir, "__dot__gitignore"), join(destdir, ".gitignore"));
 }
 
-async function copyExistingWorkerFiles(ctx: Context) {
+async function copyExistingWorkerFiles(ctx: C3Context) {
 	const { preexisting } = await getTemplate(ctx);
 
 	if (preexisting) {
@@ -166,7 +151,7 @@ async function copyExistingWorkerFiles(ctx: Context) {
 	}
 }
 
-async function updateFiles(ctx: Context) {
+async function updateFiles(ctx: C3Context) {
 	// Update package.json with project name
 	const pkgJsonPath = resolve(ctx.project.path, "package.json");
 	const pkgJson = readJSON(pkgJsonPath);
